:bug: handle error and change create review from screen to modal and fix ui and back button some page
